requirements.txt

fastapi==0.115.0
uvicorn==0.30.6
sqlmodel==0.0.22
psycopg2-binary==2.9.9
pydantic==2.11.0
pydantic-settings==2.5.2
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
alembic==1.13.2
pytest==8.3.3
httpx==0.27.2
python-multipart==0.0.12
pytest-cov==5.0.0  
pytest-cov==5.0.0